requirements.txt
Flask
Flask-Cors
python-dotenv
spacy
pandas
pymongo>=4.12.0
rapidfuzz
requests
Werkzeug
gunicorn
openpyxl
faiss-cpu
transformers
sentencepiece
sentence_transformers
importlib